introduce proper integration tests
We introduce a new kind of integraiton test where
the application performs the full flow starting from
serialized blockchain data, instead of inlined bytes
as was done in a test we replace here.

To power this, we source in some testdata that was generated
by syncing a node up to some small height. For more details,
see the new README file.