add initial block analysis to sqlite db
This commit introduces the new approach we
take towards local data analysis: build up
a database of pertinent information.

We start with some low-hanging fruit on
individual blocks.

The new DB logic supports sqlite for now,
both in in-memory mode for tests, and to
a DB file for persistence. No attempts are
made at recognizing existing data in an
existing DB.

As an optimization, we buffer 1000 blocks
before we commit data to the DB. This is
done in an admittedly hacky manner, to be
cleaned up later.