callbacks: provide access to block headers
For some callbacks access to only the header
may be sufficient, at which point deserializing
the whole block would be wasteful. We thus
provide a new hook that gets called with the
header.

If the `on_block()` hook remains unused, the
`block` variable being deserialized should be
optimized away by the compiler.