remove callbacks
We want to migrate this crate into a direction
where we provide out-of-the-box all of the
ways in which a header and / or block could
be analyzed. The user can then switch these
analyses on or off. For this functionality,
we don't need the overkill flexibility of
trait objects.

The goal is to provide a much more authoritative
block analysis tool, closer to an enhanced block
explorer than the ad-hoc analyses implemented before.