tests: replicate data instead of synchronizing access
In the future we also want to integration-test
the higher-level `BlockchainParser` struct, which
needs access to a proper `ChainStorage`. Our current
approach of synchronizing access to a single one
would lead to problems there, as it is only
available behind a mutex guard.

To solve this, we replicate the test data into a
new tempdir for every test, which then has exclusive
access. This also gets rid of the awkward Git
cleanup afterwards.